requirements.txt
asgiref==3.4.1
Django==3.2.8
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
pytz==2021.3
sqlparse==0.4.2
